pyproject.toml
[project]
name = "gemini-stock"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "feedparser>=6.0.12",
    "google-genai>=1.64.0",
    "pandas>=3.0.1",
    "pandas-ta>=0.4.71b0",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "yfinance>=1.2.0",
]
